24_07_01 10:30 음성점수 및 총점수 평균가져오기

+++ src/main/java/io/paioneer/nain/interview/model/dto/InterviewDto.java
@@ -32,7 +32,7 @@ public class InterviewDto {
 
     private double videoScore;
 
-    private int voiceScore;
+    private double voiceScore;
 
     private Date itvDate;
     private String itvDateInfo; // Date 대신 String 타입으로 변경
